requirements.txt
bittensor==9.0.*
scikit-learn==1.6.*
joblib==1.4.*
wandb==0.19.*
substrate-interface==1.7.11
pillow==10.0.*
numpy==2.0.*
aiofiles==23.2.*
torch==2.5.*
torchvision==0.*
lpips==0.1.*
pytorch-msssim==0.2.1
pydantic>=2.6,<3
huggingface-hub==0.20.*
certifi==2024.2.*
imageio==2.31.*
opencv-python>=4.5.0
rawpy
datasets